Messages and Error Checking.
Messages now stored in the information. Errors now use throws to reduce the overall check size. Tried to leave the tagLog open to various inputs. For instance, custom message objects, append text, and so on. It was a bit odd throwing an error that would both reference a message's data and provide the append...